- Added import of TextPlot settings from Origin files
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@1054 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/plot2D/AddWidgetTool.cpp
@@ -220,7 +220,7 @@ bool AddWidgetTool::eventFilter(QObject *obj, QEvent *event)
 						d_fw->setFramePen(app->d_frame_widget_pen);
 					else {
 						QPen pen = app->d_frame_widget_pen;
-						pen.setWidthF(ceil(pen.width()));
+						pen.setWidthF(ceil((double)pen.width()));
 						d_fw->setFramePen(pen);
 					}
